pyproject.toml
[project]
name = "pmap-les-faroe-eowyen-20250124"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5netcdf>=1.6.4",
    "matplotlib>=3.10.7",
    "netcdf4>=1.7.3",
    "pydantic>=2.12.3",
    "pyproj>=3.7.2",
    "pyyaml>=6.0.3",
    "rasterio>=1.4.3",
    "scipy>=1.16.2",
    "xarray>=2025.10.1",
]
